requirements.txt
git+git://github.com/dbca-wa/dpaw-utils@minimal#egg=dpaw-utils
psycopg2-binary==2.7.4
Django==2.0.4
django-confy==1.0.4
django-extensions==2.0.6
gunicorn==19.8.1
django-reversion==2.0.13
django-import-export==1.0.1
django-queryset-csv==1.0.1
xlsxwriter==1.0.5
whitenoise==3.3.1
